+++ CHANGELOG.md
@@ -5,12 +5,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.16] - 2020-03-21
+
+### Added
+
+- Added support for futures endpoints on `HTTP` (using `isdigit` to detect futures symbols).
+
+### Modified
+
+- Using `get` method for `dict` for symbol check instead of calling by key.
+- `endpoint` on `HTTP` will now default to https://api.bybit.com if no argument.
+- `retry_codes` is now user-definable in the `HTTP` arguments.
+- All logging is now on `DEBUG` level—user will need to manually set `logging_level` to `DEBUG`.
+
+## [1.1.15] - 2020-02-08
+
+### Modified
+
+- Added attempted request to `FailedRequestError` and `InvalidRequestError` for improved error logging.
+
 ## [1.1.14] - 2020-02-06
 
 ### Modified
 
 - Fixed unexpected `tuple` generation for status codes on `InvalidRequestError` and
   `FailedRequestError`.
+- Fixed how `Websocket` handles incoming `orderBook` data due to Bybit's topic naming changes.
 
 ## [1.1.13] - 2020-02-01